Added code written in the second lab
In this lab were expected to create a physical file. A physical file's
source code is essentially a blueprint for a table in a Relational
Database. When compiled, it will create that table. We can enter data
into the table using the data file utility (DFU) using the UPDDTA
command.

We were also expected to create a simple fixed-form RPGLE program.
It was simply meant to show us how unqualified file names work in IBM i.