Added week 7's lecture notes
In this lecture we learned how to perform DML statements on a master
file (a database file/table). We used a transaction file as
instructions on how to modify the master file. We ensured all
operations were logged by storing the master file in a collection which
sports automatic journaling.