Added notes on chain operations supplied in class
These notes give a more in depth explanation of the CHAIN operation. It
is not required to be read however, you only need to understand the use
of the CHAIN operation in the lab 7 code, and nothing more. As long as
you understand the code, you can save yourself the time from reading
that document.