Added week 3's lecture notes
In this week, we learned about the Data Description Specification
language (DDS). This language is used to define database tables, display
files (also known as work station files), printer/spooled files etc.
Display files are used for interactive applications. They specify the
blueprint of a screen. They consist of multiple records. Each record
individually can be read from and written to, this allows information to
only be shown if a specified condition has been met. Every record
consists of named fields (which are used as variables in RPGLE) and
constants. All named fields, constants, as well as records themselves
can have their properties manipulated by the language.

We also learned about free-format Report Program Generator Language
environment (RPGLE). The most important concept we learned is how to
transfer data between a program and display file (that is, to a
specific record in that display file):

- dcl-f -> declare file, the RPGLE program must declare the fact the we
are using the display file in our program

- Write -> transfer data to a display file

- Exfmt -> transfer data from a display file to a program