Added week 4's lecture notes
In this week, we learned about the CL programming language.
Specifically, we learned how to create a CL program to be executed
as a batch job (background job). When compiled, it creates an
Integrated Language Environment (ILE) module. For the lab, we created a
startup CL program that performs useful operations for us upon
logging in (calling STRJOB and adding BCI433LIB to my library list).

Futhermore, we have learned about User profiles, classes as
well as special authorities. You may read the user profiles supplement
reading, however, I do not recall ever needing to retain any information
on the topic, read it if your curious. Remembering what you learned in the
weekly slides and the lab about user profiles should be enough.

Note: The second part of the lab, dicussed in lecture 4.5, pertaining
to CL application programming (Parts E, F, and G, AKA. lab 4-2) was
not required to be done in the 2020 summer semester. Regardless, we
were still expected to understand CL programming as we were tested on
these concepts on the midterm.