Added source code for lab 8 challenge lab
In lab 8, we were expected to create an application that used a subfile
to display Toronto Blue Jay players to the screen. The displayed players
can be filtered by predefined filters defined on the main menu screen. A
subfile allows for showing players on page full at a time (similar to
that of a pager on unix-like systems). Subfiles allow you to write to
the same record format name multiple times (storing the data written to
the subfile in memory). When all the records are written, they are sent
to the display in one operation.