Added source code written in lab9a
In this lab, we were expected to create an application that displays
customer information from joining 2 tables (A customer table and
contacts table). We read the customer table as a keyed file and the
contacts table with SQL using SELECT ... INTO ... statements (SELECT
statements in procedural languages store there result in variables
supplied in the INTO clause). For every read operation of the customer table,
we query for the corresponding record in the contacts table (searching
for the foriegn key). This had to be done in this way as SELECT ... INTO
... statements can only return one row, any other amount is an error.
handling errors was also a requirement. We needed to watch for:

* Customers may not have a corresponding record in the contacts table.
(must check the SQL communications area)

* Fields may be null (must use an indicator)