Added Week 10's example code
The first example program shows how to properly embedded an SQL cursor
statement in a RPGLE program. It is a boilerplate program. The second
example program shows a common error (cursor not closed properly) with
students and explains the weird behaviour of the program with subsequent runs.
The third program shows the fact that you can include the ORDER BY
clause in the DECLARE CURSOR statament.