Added Lab 10's source code
This was code written for lab 10. This lab was shortened and made very
simple which taught us how to define and use user-defined functions.
In this lab we were expected to write a file that contained
prototypes and have that file be copied in anytime it is used (or
defined). To show the convience of user-defined functions, we made two
definitions for the same function and compiled the program twice with
one of the two definitions (which is automated in runlab10.clle). No
changes needed to be made to the client code across the two versions
of the application.