Added Final Research/Reflection Assignment
At the end of the semester, we were expected to create a reflection that
summarizes what was learned in a specific lab of the student's choice. I
chose to discuss about lab 8. We were also expected to research answers
to some questions. All information that I have used creating this
document is in the citations section.