Added general notes for week 9's lecture.
In week 9, we have learned how to embed SQL code (more specifically, SQL
SELECT INTO statements) inside of a RPGLE program. When you embed a
SELECT statement in a procedural language, you must include an INTO
clause which specifies where to store the retrieved data (specifically,
it specifies which host variables to store into). Using SQL is an
alternative way to retrieve data from a table (data file). However, SQL
allows for more flexibility, in that it allows for calculated
columns. Just like how the READ operation was used in lab 7, SELECT
statements can fail, and we must check for this. We do this using the
SQL communications area as well as indicator variables.