Added notes on example code supplied
Example code was supplied to assist students in this lab. The examples
show why it is important to check the SQL communications area when it is
nessecary to do so. They show the difference between including error
checking in your code (CUSTLOOKUP2) and not including error checking in
your code (CUSTLOOKUP).