Added source code for lab 7
The main idea of this lab was that we made an application that performs
DML operations on a master file (that is, a table containing all of our
data) based on instructions given by another table which we refer to as
a transaction table. With the use of the CHAIN operation, we can
acurrately determine if a requested operation in a transaction file is
actually valid.